+++ streamlit_openai/__init__.py
@@ -1,2 +1,2 @@
-from .utils import CustomFunction
+from .utils import CustomFunction, RemoteMCP
 from .chat import Chat