+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.2 (in development)
 * Add support for OpenAI's web search capability.
+* Add support for OpenAI's remote MCP capability.
 
 ## 0.1.1 (2025-06-15)
 * Fix a typo in `Chat.save`.